[refactor] Silence some of the healthcheck noise.
