+++ lib/middlewares/index.js
@@ -2,12 +2,6 @@ const cookieParser = require('cookie-parser');
 const bodyParser = require('body-parser');
 const cors = require('access-control');
 const debugParam = require('./debug');
-const debra = require('debra');
-const json = require('morgan-json');
-
-const healthcheck = /healthcheck/;
-let rid = 0;
-
 
 module.exports = function (app, options, done) {
   //
@@ -27,16 +21,6 @@ module.exports = function (app, options, done) {
   // Enable CORS based on configuration.
   //
   app.use(cors(app.config.get('cors')));
-
-  app.use(function httpLogger(req, res, next) {
-    // Reduce healthcheck logs as it creates unnecessary noise
-    if (healthcheck.test(req.url) && rid++ % 50 !== 0) return next();
-    req.log.info('Start %s - %s', req.method, req.url, {
-      method: req.method,
-      url: req.url
-    });
-    next();
-  });
 
   app.use(cookieParser());
 
@@ -63,42 +47,7 @@ module.exports = function (app, options, done) {
     }
   }));
 
-  debra.token('env', function (req) {
-    return req.params.env;
-  });
-
-  debra.token('pkg', function (req) {
-    return req.params.pkg;
-  });
-
-  debra.token('version', function (req) {
-    return req.params.version;
-  });
-
-  debra.token('hash', function (req) {
-    return req.params.hash;
-  });
-
-  var format = json({
-    'request': ':method :url HTTP/:http-version :status',
-    'remote': ':remote-addr - :remote-user',
-    'date': ':date[clf]',
-    'referrer': ':referrer',
-    'user-agent': ':user-agent',
-    'env': ':env',
-    'pkg': ':pkg',
-    'version': ':version',
-    'hash': ':hash'
-  }, { stringify: false });
-
-  app.use(debra(format, {
-    objectMode: true,
-    stream: {
-      write: function (meta) {
-        app.log.info('Request served for wrhs', meta);
-      }
-    }
-  }));
+  app.use(require('./debra')(app));
 
   //
   // After actions have been defined, we can add
